requirements.txt
selenium==4.19.0
python-dotenv==1.1.0
requests